+++ src/main/lombok/org/heigit/ohsome/ohsomeapi/inputprocessing/InputProcessor.java
@@ -68,6 +68,7 @@ public class InputProcessor {
   private Map<String, String[]> requestParameters;
   private boolean includeTags;
   private boolean includeOSMMetadata;
+  private boolean includeContributionTypes;
   private boolean clipGeometry = true;
 
   public InputProcessor(ProcessingData processingData) {
@@ -391,17 +392,15 @@ public class InputProcessor {
   }
 
   /**
-   * Processes the properties parameter used in data-extraction ressources and sets the respective
-   * boolean values includeTags, includeOSMMetadata and unclippedGeometries.
+   * Processes the properties parameter used in data-extraction resources and sets the respective
+   * boolean values includeTags, includeOSMMetadata, unclippedGeometries, and
+   * includeContributionTypes (only for the /contributions endpoints).
    * 
    * @throws BadRequestException if the properties parameter contains invalid values
    */
   public void processPropertiesParam() {
     String[] properties =
         splitParamOnComma(createEmptyArrayIfNull(requestParameters.get("properties")));
-    if (properties.length > 3) {
-      throw new BadRequestException(ExceptionMessages.PROPERTIES_PARAM);
-    }
     for (String property : properties) {
       @Deprecated
       boolean oldUnclippedParameter = "unclipped".equalsIgnoreCase(property);
@@ -411,6 +410,12 @@ public class InputProcessor {
         this.includeOSMMetadata = true;
       } else if (oldUnclippedParameter) {
         this.clipGeometry = false;
+      } else if (RequestUtils.isContributionsExtraction(requestUrl)) {
+        if ("contributionTypes".equalsIgnoreCase(property)) {
+          this.includeContributionTypes = true;
+        } else {
+          throw new BadRequestException(ExceptionMessages.PROPERTIES_PARAM_CONTR);
+        }
       } else {
         throw new BadRequestException(ExceptionMessages.PROPERTIES_PARAM);
       }
@@ -828,6 +833,10 @@ public class InputProcessor {
     return includeOSMMetadata;
   }
 
+  public boolean includeContributionTypes() {
+    return includeContributionTypes;
+  }
+  
   public boolean isClipGeometry() {
     return clipGeometry;
   }